requirements.txt
pyyaml
duckdb
pandas